Fix investment shares type mismatch (Bug #13)
Changed HoldingSnapshot.quantity and HoldingPerformance.quantity from
Long (in 1/10000 units) to Double to be consistent with Holding.shares
and HoldingLot.shares which already use Double.

This simplifies the codebase by eliminating the need for x10000 and
/10000 conversions and ensures consistent precision across all models.

Updated:
- Performance.kt: HoldingSnapshot.quantity and HoldingPerformance.quantity
- PerformanceRepositoryImpl.kt: Removed conversion multiplications
- PerformanceTabContent.kt: formatShares now takes Double directly
- HoldingDetailScreen.kt: formatShares now takes Double directly
- HoldingDetailViewModelTest.kt: Updated test data

+++ shared/src/commonMain/kotlin/com/financeapp/domain/model/Performance.kt
@@ -71,7 +71,7 @@ data class HoldingSnapshot(
     val id: Long = 0,
     val holdingId: Long,
     val date: Long,
-    val quantity: Long, // In 1/10000 shares
+    val quantity: Double, // Number of shares (consistent with Holding.shares)
     val price: Long, // In cents
     val value: Long, // In cents
     val costBasis: Long, // In cents
@@ -86,7 +86,7 @@ data class HoldingPerformance(
     val holdingId: Long,
     val symbol: String,
     val name: String,
-    val quantity: Long, // In 1/10000 shares
+    val quantity: Double, // Number of shares (consistent with Holding.shares)
     val costBasis: Long, // Total cost basis in cents
     val currentPrice: Long, // Current price in cents
     val currentValue: Long, // Current total value in cents
